Finished remote OT network testing automation
Finished and tested the new way of automating the set up, test
and measurement of an arbitrary OpenThread network.
Fixed some small bugs around the set up, including logic errors
in conditions, making sure the correct nodes are used, etc.
Prepared two sets of `nodes` and `topology` YAML files, one for
a network of two and one for a network of four nodes.
Tested everything with a two-node network in the lab.